requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.1
colorama==0.4.6
frozenlist==1.4.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.4
openai==0.28.1
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
tqdm==4.66.1
urllib3==2.0.7
yarl==1.9.2
